5_2_4: important changes, with part of the following task. I still need to figure with squash commits .. and so on, does not work.. rly, it's not joke, i use git rebase -i HEAD~n, use :wq, but result bad

+++ src/main/java/org/ot5usk/entities/Book.java
@@ -2,7 +2,9 @@ package org.ot5usk.entities;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.sql.Timestamp;
 
@@ -15,7 +17,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;
+    private Integer id;
 
     @Column(name = "book_title")
     private String bookTitle;
